+++ mrb_doc/core/window.rb
@@ -7,24 +7,6 @@ def set_window_monitor(monitor)
   nil
 end
 
-# Gets the width of the specified monitor.
-# @param monitor [Integer]
-# @return [Integer]
-def get_monitor_width(monitor)
-  # mrb_get_monitor_width
-  # src/mruby_integration/core/window.cpp
-  1920
-end
-
-# Gets the height of the specified monitor.
-# @param monitor [Integer]
-# @return [Integer]
-def get_monitor_height(monitor)
-  # mrb_get_monitor_height
-  # src/mruby_integration/core/window.cpp
-  1080
-end
-
 # Gets the refresh rate of the specified monitor.
 # @param monitor [Integer]
 # @return [Integer]