+++ src/mruby_integration/models/monitor.cpp
@@ -96,6 +96,32 @@ mrb_Monitor_position(mrb_state* mrb, mrb_value self) -> mrb_value
   return obj;
 }
 
+auto
+mrb_Monitor_width(mrb_state* mrb, mrb_value self) -> mrb_value
+{
+  EXIT_UNLESS_WINDOW_READY("You must call Window.open before Monitor#width");
+
+  Monitor* monitor;
+
+  Data_Get_Struct(mrb, self, &Monitor_type, monitor);
+  mrb_assert(monitor != nullptr);
+
+  return mrb_int_value(mrb, GetMonitorWidth(monitor->id));
+}
+
+auto
+mrb_Monitor_height(mrb_state* mrb, mrb_value self) -> mrb_value
+{
+  EXIT_UNLESS_WINDOW_READY("You must call Window.open before Monitor#height");
+
+  Monitor* monitor;
+
+  Data_Get_Struct(mrb, self, &Monitor_type, monitor);
+  mrb_assert(monitor != nullptr);
+
+  return mrb_int_value(mrb, GetMonitorHeight(monitor->id));
+}
+
 void
 append_models_Monitor(mrb_state* mrb)
 {
@@ -111,6 +137,10 @@ append_models_Monitor(mrb_state* mrb)
     mrb, Monitor_class, "initialize", mrb_Monitor_initialize, MRB_ARGS_REQ(1));
   mrb_define_method(
     mrb, Monitor_class, "position", mrb_Monitor_position, MRB_ARGS_NONE());
+  mrb_define_method(
+    mrb, Monitor_class, "width", mrb_Monitor_width, MRB_ARGS_NONE());
+  mrb_define_method(
+    mrb, Monitor_class, "height", mrb_Monitor_height, MRB_ARGS_NONE());
 
   load_ruby_models_monitor(mrb);
 }